+++ limix/_cli/_scan/__init__.py
@@ -1,0 +1,3 @@
+from ._scan import scan
+
+__all__ = ["scan"]
